pyproject.toml
[project]
name = "ja-pagou"
version = "0.1.0"
description = ""
authors = [
    {name = "Arthur Fernandes",email = "arthur0139@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "django (==4.2)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "markdown (>=3.7,<4.0)",
    "django-filter (>=24.3,<25.0)",
    "channels (>=4.2.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.1"
factory-boy = "^3.3.1"
faker = "^35.0.0"

[tool.taskipy.tasks]
run = { cmd = "python manage.py runserver", help = "Running Django" }
makemigrations = { cmd = "python manage.py makemigrations", help = "Makemigrations Django" }
migrate = { cmd = "python manage.py migrate", help = "Migrate Django" }
seed = { cmd = "python manage.py seed", help = "Seeding Django" }

